Support for AUR and swap

+++ start.sh
@@ -8,7 +8,10 @@ done < "packages.install"
 
 read -p "boot partition name: " boot_part
 read -p "root partition name: " root_part
-read -p "home partition name: " home_part
+read -p "with home? [y/N]: " has_home
+if [ "$has_home" = "y" ]; then read -p "home partition name: " home_part; fi
+read -p "with swap? [y/N]: " has_swap
+if [ "$has_swap" = "y" ]; then read -p "swap partition name: " swap_part; fi
 read -sp "root password: " root_pass
 echo " "
 read -p "user name: " user
@@ -22,7 +25,7 @@ read -p "hostname: " hostname
 timedatectl set-ntp true
 
 mkfs.ext4 "$root_part"
-mkfs.ext4 "$home_part"
+if [ "$has_home" = "y" ]; then mkfs.ext4 "$home_part"; fi
 mkfs.vfat -F 32 "$boot_part"
 
 mount "$root_part" /mnt
@@ -31,8 +34,17 @@ mkdir /mnt/boot
 mkdir /mnt/boot/EFI
 mount "$boot_part" /mnt/boot/EFI
 
-mkdir /mnt/home
-mount "$home_part" /mnt/home
+if [ "$has_home" = "y" ]
+then 
+	mkdir /mnt/home
+	mount "$home_part" /mnt/home
+fi
+
+if [ "$has_swap" = "y" ]
+then 
+	mkswap "$swap_part"
+	swapon "$swap_part"
+fi
 
 pacstrap /mnt ${pkg[*]} grub efibootmgr # installs all required packages
 
@@ -46,5 +58,5 @@ rm /mnt/install.sh
 rm /mnt/aur.install
 rm /mnt/services.install
 
-umount -R /mnt
-reboot
+#umount -R /mnt
+#reboot
